CM-208: refactored interface approach
