--- calcrule_social_protection/interfaces/__init__.py
@@ -1,2 +1,0 @@
-from calcrule_social_protection.interfaces.group_interface import GroupBenefitPlanInterface
-from calcrule_social_protection.interfaces.individual_interface import IndividualBenefitPlanInterface
