+++ calcrule_social_protection/strategies/__init__.py
@@ -1,0 +1,5 @@
+from calcrule_social_protection.strategies.benefit_package_base_strategy import BaseBenefitPackageStrategy
+from calcrule_social_protection.strategies.benefit_package_strategy_interface import BenefitPackageStrategyInterface
+from calcrule_social_protection.strategies.benefit_package_strategy_storage import BenefitPackageStrategyStorage
+from calcrule_social_protection.strategies.benefit_package_group_strategy import GroupBenefitPackageStrategy
+from calcrule_social_protection.strategies.benefit_package_individual_strategy import IndividualBenefitPackageStrategy
